+++ tests/executor_test.py
@@ -6,12 +6,13 @@ from tango.executor import ExecutorMetadata
 class TestMetadata(TangoTestCase):
     def test_metadata(self):
         metadata = ExecutorMetadata("some_step")
-        metadata.finalize()
+        metadata.save(self.TEST_DIR)
 
         assert metadata.git is not None
         assert metadata.git.remote is not None
         assert "allenai/tango" in metadata.git.remote
 
-        metadata.to_params().to_file(self.TEST_DIR / "metadata.json")
-        metadata2 = ExecutorMetadata.from_params(Params.from_file(self.TEST_DIR / "metadata.json"))
+        metadata2 = ExecutorMetadata.from_params(
+            Params.from_file(self.TEST_DIR / "executor-metadata.json")
+        )
         assert metadata == metadata2