+++ tests/main_test.py
@@ -2,7 +2,9 @@ import os
 import subprocess
 
 from tango.version import VERSION
+from tango.common import Params
 from tango.common.testing import TangoTestCase
+from tango.executor import ExecutorMetadata
 
 
 class TestMain(TangoTestCase):
@@ -28,7 +30,22 @@ class TestMain(TangoTestCase):
         assert (self.TEST_DIR / "hello" / "cache-metadata.json").is_file()
         assert (self.TEST_DIR / "hello" / "executor-metadata.json").is_file()
         assert (self.TEST_DIR / "hello_world").is_dir()
-        assert (self.TEST_DIR / "hello_world" / "executor-metadata.json").is_file()
+
+        # Check metadata.
+        metadata_path = self.TEST_DIR / "hello_world" / "executor-metadata.json"
+        assert metadata_path.is_file()
+        metadata_params = Params.from_file(metadata_path)
+        metadata = ExecutorMetadata.from_params(metadata_params)
+        assert metadata.config == {
+            "type": "concat_strings",
+            "string1": {"type": "ref", "ref": "hello"},
+            "string2": "World!",
+            "join_with": ", ",
+        }
+        assert metadata.git.commit is not None
+        assert metadata.git.remote is not None
+
+        assert (self.TEST_DIR / "hello_world" / "requirements.txt").is_file()
 
         # Running again shouldn't create any more directories.
         result = subprocess.run(cmd)